img-worker/pyproject.toml
[project]
name = "pixelforge-img-worker"
version = "0.1.0"
description = ""
authors = [
    {name = "hfidelis",email = "heitorc88@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery (>=5.5.3,<6.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
